General cleanup of Kazari code
